+++ agent/agent-profiler/agent-service-profiler/src/main/java/com/microsoft/applicationinsights/serviceprofilerapi/util/TimestampContract.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.applicationinsights.serviceprofilerapi.client.contract;
+package com.microsoft.applicationinsights.serviceprofilerapi.util;
 
 import java.time.Instant;
 import java.time.OffsetDateTime;
